requirements.txt
{\rtf1\ansi\ansicpg1252\cocoartf2822
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica;}
{\colortbl;\red255\green255\blue255;}
{\*\expandedcolortbl;;}
\paperw11900\paperh16840\margl1440\margr1440\vieww11520\viewh8400\viewkind0
\pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\pardirnatural\partightenfactor0

\f0\fs24 \cf0 # Core\
python-dotenv\
requests\
jsonschema\
\
# OpenAI\
openai>=1.0.0\
\
# Hugging Face + Training\
torch>=2.2.0\
transformers>=4.41.0\
datasets>=2.20.0\
accelerate>=0.30.0\
trl>=0.9.6\
peft>=0.11.1\
bitsandbytes>=0.43.0\
\
# Utilities\
tqdm\
}